chore(client): Isolate FPVM-specific constructs
