+++ bin/client/src/fault/mod.rs
@@ -1,10 +1,13 @@
-//! Contains the host <-> client communication utilities.
+//! Contains FPVM-specific constructs for the `kona-client` program.
 
 use kona_common::FileDescriptor;
 use kona_preimage::{HintWriter, OracleReader, PipeHandle};
 
 mod caching_oracle;
-pub use caching_oracle::CachingOracle;
+pub(crate) use caching_oracle::CachingOracle;
+
+mod precompiles;
+pub(crate) use precompiles::FPVMPrecompileOverride;
 
 /// The global preimage oracle reader pipe.
 static ORACLE_READER_PIPE: PipeHandle =
@@ -15,7 +18,7 @@ static HINT_WRITER_PIPE: PipeHandle =
     PipeHandle::new(FileDescriptor::HintRead, FileDescriptor::HintWrite);
 
 /// The global preimage oracle reader.
-pub static ORACLE_READER: OracleReader = OracleReader::new(ORACLE_READER_PIPE);
+pub(crate) static ORACLE_READER: OracleReader = OracleReader::new(ORACLE_READER_PIPE);
 
 /// The global hint writer.
-pub static HINT_WRITER: HintWriter = HintWriter::new(HINT_WRITER_PIPE);
+pub(crate) static HINT_WRITER: HintWriter = HintWriter::new(HINT_WRITER_PIPE);
