+++ bin/client/src/fault/precompiles/bn128_pair.rs
@@ -1,36 +1,41 @@
-//! Contains the accelerated version of the KZG point evaluation precompile.
+//! Contains the accelerated version of the `ecPairing` precompile.
 
+use crate::fault::{HINT_WRITER, ORACLE_READER};
 use alloc::{string::ToString, vec::Vec};
 use alloy_primitives::{keccak256, Address, Bytes};
 use anyhow::ensure;
+use kona_client::HintType;
 use kona_preimage::{HintWriterClient, PreimageKey, PreimageKeyType, PreimageOracleClient};
 use revm::{
-    precompile::{u64_to_address, Error as PrecompileError, PrecompileWithAddress},
+    precompile::{
+        bn128::pair::{ISTANBUL_PAIR_BASE, ISTANBUL_PAIR_PER_POINT},
+        u64_to_address, Error as PrecompileError, PrecompileWithAddress,
+    },
     primitives::{Precompile, PrecompileOutput, PrecompileResult},
 };
 
-use crate::{HintType, HINT_WRITER, ORACLE_READER};
+const ECPAIRING_ADDRESS: Address = u64_to_address(8);
+const PAIR_ELEMENT_LEN: usize = 64 + 128;
 
-const POINT_EVAL_ADDRESS: Address = u64_to_address(0x0A);
+pub(crate) const FPVM_ECPAIRING: PrecompileWithAddress =
+    PrecompileWithAddress(ECPAIRING_ADDRESS, Precompile::Standard(fpvm_ecpairing));
 
-pub(crate) const FPVM_KZG_POINT_EVAL: PrecompileWithAddress =
-    PrecompileWithAddress(POINT_EVAL_ADDRESS, Precompile::Standard(fpvm_kzg_point_eval));
+/// Performs an FPVM-accelerated `ecpairing` precompile call.
+fn fpvm_ecpairing(input: &Bytes, gas_limit: u64) -> PrecompileResult {
+    let gas_used =
+        (input.len() / PAIR_ELEMENT_LEN) as u64 * ISTANBUL_PAIR_PER_POINT + ISTANBUL_PAIR_BASE;
 
-/// Performs an FPVM-accelerated KZG point evaluation precompile call.
-fn fpvm_kzg_point_eval(input: &Bytes, gas_limit: u64) -> PrecompileResult {
-    const GAS_COST: u64 = 50_000;
-
-    if gas_limit < GAS_COST {
+    if gas_used > gas_limit {
         return Err(PrecompileError::OutOfGas.into());
     }
 
-    if input.len() != 192 {
-        return Err(PrecompileError::BlobInvalidInputLength.into());
+    if input.len() % PAIR_ELEMENT_LEN != 0 {
+        return Err(PrecompileError::Bn128PairLength.into());
     }
 
     let result_data = kona_common::block_on(async move {
         // Write the hint for the ecrecover precompile run.
-        let hint_data = &[POINT_EVAL_ADDRESS.as_ref(), input.as_ref()];
+        let hint_data = &[ECPAIRING_ADDRESS.as_ref(), input.as_ref()];
         HINT_WRITER.write(&HintType::L1Precompile.encode_with(hint_data)).await?;
 
         // Construct the key hash for the ecrecover precompile run.
@@ -52,5 +57,5 @@ fn fpvm_kzg_point_eval(input: &Bytes, gas_limit: u64) -> PrecompileResult {
     })
     .map_err(|e| PrecompileError::Other(e.to_string()))?;
 
-    Ok(PrecompileOutput::new(GAS_COST, result_data.into()))
+    Ok(PrecompileOutput::new(gas_used, result_data.into()))
 }
