+++ bin/client/src/l2/mod.rs
@@ -2,6 +2,3 @@
 
 mod chain_provider;
 pub use chain_provider::OracleL2ChainProvider;
-
-mod precompiles;
-pub use precompiles::FPVMPrecompileOverride;
